Convert 'Bugzilla' references to the standard format
In virt team, contributors use

    Bugzilla: <bzid OR bzlink>

in the MR message to reference bugs in Red Hat's Bugzilla.

This format, though, is different from the one required by BZ checks in
dist-git commit messages, which is (among others):

    Resolves: bz#<bzid>

Do a multiline substitution, to convert the former to the later, and use
this string as the message for the dist-git MR and the commit message.

Signed-off-by: Hunor Csomortáni <csomh@redhat.com>