Add a log message to debug missing merge_request_url
to know whether is was a flake or there's some pattern.
https://github.com/packit/packit-service/blob/da4a797985708b814e744a7c19610f00384d6729/packit_service/worker/parser.py#L1193
says that the merge_request can be missing in the original
webhook payload if the 'source' is 'push', but in the
https://sentry.io/organizations/red-hat-0p/issues/3427461150
case the 'source' was 'merge_request_event'.

The trigger was
https://gitlab.com/redhat/centos-stream/rpms/gnome-settings-daemon/-/pipelines/588479707
from
https://gitlab.com/redhat/centos-stream/rpms/gnome-settings-daemon/-/merge_requests/7
which doesn't look anyhow special.

Unfortunatelly, we don't know what was in the payload,
https://github.com/packit/packit-service/issues/1512
so let's add some debug logging to have some clues next time.